# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WellPath API",
    "version": "0.1.0",
    "description": "Health assessment subscription funnel. Errors follow RFC 9457 (application/problem+json). Access is bound to an HttpOnly same-site cookie issued at session creation."
  },
  "servers": [{ "url": "https://{host}", "variables": { "host": { "default": "localhost:3000" } } } ],
  "tags": [
    { "name": "system" },
    { "name": "assessment" },
    { "name": "billing" }
  ],
  "paths": {
    "/api/health": {
      "get": {
        "tags": ["system"],
        "summary": "Liveness + DB ping (also used by cron keep-alive)",
        "responses": {
          "200": {
            "description": "OK",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "status": { "const": "ok" }, "db": { "enum": ["up", "down"] }, "version": { "type": "string" } }, "required": ["status", "db"] } } }
          }
        }
      }
    },
    "/api/assessments": {
      "post": {
        "tags": ["assessment"],
        "summary": "Create a session, free entitlement and 24h access cookie",
        "responses": {
          "201": { "description": "Created; Set-Cookie: wellpath_sid", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionCreated" } } } },
          "429": { "$ref": "#/components/responses/RateLimited" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
    "/api/assessments/{id}": {
      "get": {
        "tags": ["assessment"],
        "summary": "Resume a session (steps + revisions)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } },
          "400": { "$ref": "#/components/responses/Invalid" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["assessment"],
        "summary": "Hard delete a session and all cascaded rows (GDPR-style erasure)",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Deleted" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/assessments/{id}/steps/{stepKey}": {
      "patch": {
        "tags": ["assessment"],
        "summary": "Upsert one step answer with optimistic locking",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "stepKey", "in": "path", "required": true, "schema": { "enum": ["basics", "goal", "activity", "condition"] } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StepPatch" } } } },
        "responses": {
          "200": { "description": "Saved", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StepSaved" } } } },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/assessments/{id}/submit": {
      "post": {
        "tags": ["assessment"],
        "summary": "Atomically assemble, validate, compute and persist the result. Repeated submit is idempotent; recalculate=true recomputes after editing.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "requestBody": { "content": { "application/json": { "schema": { "type": "object", "properties": { "clientToken": { "type": "string", "format": "uuid" }, "recalculate": { "type": "boolean", "default": false } } } } } },
        "responses": {
          "200": { "description": "Submitted", "content": { "application/json": { "schema": { "type": "object", "properties": { "sessionId": { "type": "string" }, "recomputed": { "type": "boolean" }, "kind": { "enum": ["complete", "protected"] } }, "required": ["sessionId", "recomputed", "kind"] } } } },
          "422": { "$ref": "#/components/responses/Validation" }
        }
      }
    },
    "/api/assessments/{id}/result": {
      "get": {
        "tags": ["assessment"],
        "summary": "Result DTO gated by entitlement. Free returns only freeSummary + locked=true; premium returns full payload.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Free or full DTO", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ResultView" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/assessments/{id}/insights": {
      "post": {
        "tags": ["assessment"],
        "summary": "Premium-only AI coach insights. Calls a local Ollama (DeepSeek) model with the already-computed numbers; falls back to deterministic rule tips (source=rule-fallback) when no local model is reachable. Non-members get 402.",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }],
        "responses": {
          "200": { "description": "Tips for a premium session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AiInsights" } } } },
          "402": { "description": "Premium subscription required", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
        }
      }
    },
    "/api/pay": {
      "post": {
        "tags": ["billing"],
        "summary": "Simulated payment. Idempotency-Key permanent dedupe; same key diff payload -> 409. Grants 30d premium and one-time recovery code.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PayRequest" } } } },
        "responses": {
          "200": { "description": "Paid (or replayed). recoveryCode is null on replay.", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PayResult" } } } },
          "409": { "$ref": "#/components/responses/Conflict" },
          "422": { "$ref": "#/components/responses/Validation" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/recovery/redeem": {
      "post": {
        "tags": ["billing"],
        "summary": "Redeem a single-use recovery code; issues a fresh 24h access cookie",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["recoveryCode"], "properties": { "recoveryCode": { "type": "string", "minLength": 16 } } } } } },
        "responses": {
          "200": { "description": "Redeemed; Set-Cookie" },
          "401": { "$ref": "#/components/responses/RecoveryInvalid" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "Invalid": { "description": "Bad request", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "RecoveryInvalid": { "description": "Recovery code unknown, used or expired", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "NotFound": { "description": "Session not found", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Conflict": { "description": "Optimistic-lock or idempotency conflict; body.fieldErrors.currentRevision carries server revision", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Validation": { "description": "Schema/business validation failed with fieldErrors", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "RateLimited": { "description": "Fixed-window rate exceeded; Retry-After", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
      "Internal": { "description": "Internal error (no stack leaked)", "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "properties": {
          "type": { "type": "string" },
          "title": { "type": "string" },
          "status": { "type": "integer" },
          "code": { "enum": ["INVALID_REQUEST", "RECOVERY_INVALID", "SESSION_NOT_FOUND", "STEP_CONFLICT", "PAYMENT_IDEMPOTENT_MISMATCH", "PAYMENT_REQUIRED", "VALIDATION_FAILED", "RATE_LIMITED", "INTERNAL_ERROR"] },
          "detail": { "type": "string" },
          "requestId": { "type": "string" },
          "fieldErrors": { "type": "object", "additionalProperties": { "type": "array", "items": { "type": "string" } } }
        },
        "required": ["code", "status"]
      },
      "AiInsights": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "string", "format": "uuid" },
          "access": { "const": "full" },
          "source": { "enum": ["local-llm", "rule-fallback"], "description": "local-llm=generated by the local Ollama/DeepSeek model; rule-fallback=deterministic tips because no local model was reachable" },
          "model": { "type": "string", "description": "Model name when source=local-llm" },
          "tips": { "type": "array", "items": { "type": "string" } },
          "generatedAt": { "type": "string", "format": "date-time" }
        },
        "required": ["access", "source", "tips"]
      },
      "SessionCreated": { "type": "object", "properties": { "sessionId": { "type": "string", "format": "uuid" }, "status": { "const": "in_progress" }, "accessExpiresAt": { "type": "string", "format": "date-time" } }, "required": ["sessionId", "status"] },
      "Session": { "type": "object", "properties": { "sessionId": { "type": "string", "format": "uuid" }, "status": { "enum": ["in_progress", "submitted", "deleted"] }, "steps": { "type": "array", "items": { "$ref": "#/components/schemas/StepSaved" } } }, "required": ["sessionId", "status", "steps"] },
      "StepPatch": { "type": "object", "properties": { "stepKey": { "enum": ["basics", "goal", "activity", "condition"] }, "answer": { "type": "object", "additionalProperties": true, "description": "Partial draft of that step is accepted (the UI collects one step across several screens), but every supplied value must be well-typed/in-enum; completeness is enforced at submit. basics{sex,ageYears,heightCm,weightKg,bodyBuild?}, goal{goal?,targetWeightKg,pace?}, activity{activity,dailyMovement?,workoutPreferences?[],stairTolerance?}, condition{specialCondition,weightTendency?,focusAreas?[]}" }, "expectedRevision": { "type": "integer", "minimum": 1, "description": "Omit on first write; required afterwards for optimistic locking" } }, "required": ["stepKey", "answer"] },
      "StepSaved": { "type": "object", "properties": { "stepKey": { "type": "string" }, "revision": { "type": "integer" }, "rebased": { "type": "boolean" } }, "required": ["stepKey", "revision"] },
      "ResultView": {
        "type": "object",
        "properties": {
          "sessionId": { "type": "string", "format": "uuid" },
          "access": { "enum": ["free", "full", "protected"], "description": "free=non-member (gated), full=premium, protected=safety path (no deficit plan)" },
          "entitlementTier": { "enum": ["free", "premium"], "description": "Server-side subscription status resolved from the entitlement table (expiry-aware)" },
          "locked": { "type": "boolean", "description": "true for non-members" },
          "lockedFields": {
            "type": "array",
            "description": "Field names hidden from non-members (upsell guide). Non-empty for free; empty array for full.",
            "items": { "type": "object", "properties": { "key": { "type": "string" }, "label": { "type": "string" } }, "required": ["key", "label"] }
          },
          "upgrade": { "type": "object", "description": "Present only for free; structured paywall CTA", "properties": { "required": { "type": "boolean" }, "productCode": { "type": "string" }, "endpoint": { "type": "string" }, "message": { "type": "string" } } },
          "entitlementExpiresAt": { "type": "string", "format": "date-time", "description": "Premium expiry; present only when access=full" },
          "freeSummary": { "type": "object", "properties": { "bmi": { "type": "number" }, "bmiCategory": { "type": "string" }, "isHealthyTarget": { "type": "boolean" }, "weightDeltaKg": { "type": "number" }, "goal": { "enum": ["lose", "maintain", "gain"] }, "pace": { "enum": ["steady", "moderate", "fast"] }, "energyDirection": { "enum": ["deficit", "maintenance", "surplus"] }, "targetDateRangeWeeks": { "type": "object", "nullable": true, "properties": { "fastestWeeks": { "type": "integer" }, "steadyWeeks": { "type": "integer" } } }, "headline": { "type": "string" } }, "description": "Non-sensitive overview only; never contains bmr/tdee/recommendedIntake/activityFactor values" },
          "payload": { "type": "object", "description": "Present only when access=full. result{goal,pace,energyDirection,energyAdjustment,bmi,bmr,tdee,recommendedIntake,activityFactor,weightDeltaKg,targetDateRangeWeeks,warnings}, plus profile (report-only answers) and deterministic recommendations[]." }
        },
        "required": ["access", "entitlementTier", "locked", "lockedFields"]
      },
      "PayRequest": { "type": "object", "properties": { "sessionId": { "type": "string", "format": "uuid" }, "idempotencyKey": { "type": "string", "format": "uuid" }, "productCode": { "type": "string", "default": "wellpath_premium_30d" }, "simulate": { "enum": ["fail"], "description": "Non-production failure injection" } }, "required": ["sessionId", "idempotencyKey", "productCode"] },
      "PayResult": { "type": "object", "properties": { "status": { "enum": ["succeeded", "failed"] }, "replayed": { "type": "boolean" }, "subscriptionId": { "type": "string" }, "recoveryCode": { "type": "string", "nullable": true, "description": "Returned only once, on the first successful payment" }, "premiumExpiresAt": { "type": "string", "format": "date-time", "nullable": true } }, "required": ["status", "replayed", "recoveryCode"] }
    }
  }
}